Create page listing view (and feed)
The page listing view list multiple page teasers, paged,
10 at a time, newest first.

This view was called PG2 in the Drupal 7 version of the
platform.

In addition to the teaser listing, this view also provides
an RSS feed of either all pages (pages/all/feed) or pages
filtered by term ID (pages/%/feed).

The page display is duplicated because in Drupal 8, unlike 7,
paths with arguments do not match if the argument is not
specified (e.g. the URL /pages does not match the path /pages/%).