Migrate images to media entities.
This commit replaces the image field on news with a media reference field.

A new media bundle, image, is created to store images. The image bundle uses an image field to store images locally.

The new media reference field on the news content type references a media entity of type image.

The D7 to D8 migration has been modified to create a new image entity for each news node. The news node migration maps the entity reference fields to
these new media entities, so the new image entities are attached to the original news article.

In order to edit and create new news nodes, a simple entity browser and inline entity form have been configured to allow users to upload new image media,
or select from existing image media when creating and editing news.

News images are the only files/images to be converted to media entities so far.